Refactor DataTable class to be more customizeable
Add column header definitions to DataTable class. Users can define
each column with a title, type, and list of callback information.
When a new row is added, the column definition will be used to define
each cell. The addRow function will now take an array of text that will
be used as the text content of each cell.